Add first cut at the breast cancer registry ontology and the registry bridge ontology.
LinkML didn't come close to generating what we needed, so this was created with Protege.  We've encountered a number of interesting issues:
1) "id" was intended to reference the person being diagnosed.  At the moment, the YAML is incorrect, as we have stated that it is the
identifier of the BCR record itself.  We will fix this
2) "entryDate" should map to a _predicate_ that identifies the date that the DX occurred.  Curiously, there is a _class_, "date of diagnosis" (https://www.ebi.ac.uk/ols/ontologies/efo/terms?short_form=EFO_0004953) that sort of identifies what our predicate is, but I have NO idea how we'd represent an actual date in this situation.  We need to ask the EFO folks.

Also note that we attempted to import MONDO for the breast cancer diagnosis -- the URL that we got from OLS didn't work because it had a formatting error.  We picked up a URL from BioPortal but, after 20 minutes, it was still loading so we gave up and moved to HPO